requirements.txt
-r requirements.txt
nose==1.2.1
coverage==3.6
tox==1.4.3
pytz==2013.9
django-annoying
